requirements.txt
feedparser==6.0.11
requests==2.32.3
python-dotenv==1.1.0
flask==3.1.1
gunicorn==23.0.0
openai==1.82.0
